Moved test data to tests/data. Made TextContent a native concept.
