+++ pipelex/core/concept_native.py
@@ -18,6 +18,7 @@ class NativeConceptClass(StrEnum):
     TEXT_AND_IMAGES = "TextAndImagesContent"
     NUMBER = "NumberContent"
     LLM_PROMPT = "LLMPromptContent"
+    PAGE = "PageContent"
 
 
 class NativeConceptCode(StrEnum):
@@ -28,6 +29,7 @@ class NativeConceptCode(StrEnum):
     TEXT_AND_IMAGES = "TextAndImage"
     NUMBER = "Number"
     LLM_PROMPT = "LlmPrompt"
+    PAGE = "Page"
 
     @property
     def concept_code(self) -> str:
@@ -85,6 +87,13 @@ class NativeConceptCode(StrEnum):
                     domain=SpecialDomain.NATIVE,
                     definition="A dynamic concept",
                     structure_class_name=NativeConceptClass.DYNAMIC,
+                )
+            case NativeConceptCode.PAGE:
+                return Concept(
+                    code=ConceptFactory.make_concept_code(SpecialDomain.NATIVE, code),
+                    domain=SpecialDomain.NATIVE,
+                    definition="The content of a page of a document, comprising text and linked images as well as an optional screenshot of the page",
+                    structure_class_name=NativeConceptClass.PAGE,
                 )
 
     @classmethod
