+++ pipelex/libraries/pipelines/documents.toml
@@ -7,7 +7,6 @@ definition = "The domain of documents that can comprise pages, text, images, etc
 
 [concept]
 TextAndImagesContent = "A content that comprises text and images where the text can include local links to the images"
-PageContent = "The content of a page of a document, comprising text and linked images as well as an optional screenshot of the page"
 PDF = "A PDF document"
 
 [pipe]
