+++ tests/data/documents/solar_system.png
