+++ tests/data/images/animal_lympics.png
