+++ tests/data/images/eiffel_tower.jpg
