+++ tests/data/images/eiffel_tower.png
