+++ tests/data/images/gantt_tree_house.png
