+++ tests/pipelex/asynch/test_pipe_ocr.py
@@ -7,8 +7,8 @@ import pytest
 
 from pipelex import pretty_print
 from pipelex.core.domain import SpecialDomain
+from pipelex.core.stuff_content import PageContent
 from pipelex.core.working_memory_factory import WorkingMemoryFactory
-from pipelex.libraries.pipelines.documents import PageContent
 from pipelex.pipe_operators.pipe_ocr import PipeOCR, PipeOCROutput
 from pipelex.pipe_works.pipe_job_factory import PipeJobFactory
 from pipelex.pipe_works.pipe_router_protocol import PipeRouterProtocol
