--- tests/pipelex/data/test_pipe_batch/test_pipe_batch.py
@@ -1,11 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Evotis S.A.S.
-# SPDX-License-Identifier: Elastic-2.0
-# "Pipelex" is a trademark of Evotis S.A.S.
-
-DATA_1 = "data_1"
-DATA_2 = "data_2"
-DATA_3 = "data_3"
-DATA_4 = "data_4"
-DATA_5 = "data_5"
-DATA_6 = "data_6"
-DATA_7 = "data_7"
