+++ tests/pipelex/test_data.py
@@ -297,13 +297,13 @@ class LibraryTestCases:
 class PipeOCRTestCases:
     PIPE_OCR_IMAGE_TEST_CASES: ClassVar[List[str]] = [
         # LOCAL
-        "tests/cogt/data/documents/solar_system.png",
+        "tests/data/documents/solar_system.png",
         # REMOTE
         "https://storage.googleapis.com/public_test_files_7fa6_4277_9ab/documents/solar_system.png",
     ]
     PIPE_OCR_PDF_TEST_CASES: ClassVar[List[str]] = [
         # LOCAL
-        "tests/cogt/data/documents/solar_system.pdf",
+        "tests/data/documents/solar_system.pdf",
         # REMOTE
         "https://storage.googleapis.com/public_test_files_7fa6_4277_9ab/documents/solar_system.pdf",
     ]
