+++ tests/test_data.py
@@ -4,11 +4,11 @@
 
 
 class PDFTestCases:
-    DOCUMENT_FILE_PATH = "tests/cogt/data/documents/solar_system.pdf"
+    DOCUMENT_FILE_PATH = "tests/data/documents/solar_system.pdf"
     DOCUMENT_URL = "https://storage.googleapis.com/public_test_files_7fa6_4277_9ab/documents/solar_system.pdf"
 
 
 # TODO: move the other shared test images here from other test_data.py files
 class ImageTestCases:
-    IMAGE_FILE_PATH = "tests/cogt/data/documents/solar_system.png"
+    IMAGE_FILE_PATH = "tests/data/documents/solar_system.png"
     IMAGE_URL = "https://storage.googleapis.com/public_test_files_7fa6_4277_9ab/documents/solar_system.png"
